+++ packages/react-learn-card/src/components/VCDisplayCard2/VCIDDIsplayCard.tsx
@@ -66,12 +66,10 @@ export const VCIDDisplayCard: React.FC<VCIDDisplayCardProps> = ({
     const setIsFront = setIsFrontOverride ?? _setIsFront;
 
     let backgroundStyle = {
-        backgroundColor: `linear-gradient(180deg, rgba(24,34,78,1) ${
-            !isFront && hideGradientBackground ? '100%' : '25%'
-        }, rgba(139,145,167,1) 100%)`,
-        backgroundImage: `linear-gradient(180deg, rgba(24,34,78,1) ${
-            !isFront && hideGradientBackground ? '100%' : '25%'
-        }, rgba(139,145,167,1) 100%)`,
+        backgroundColor: `linear-gradient(180deg, rgba(24,34,78,1) ${!isFront && hideGradientBackground ? '100%' : '25%'
+            }, rgba(139,145,167,1) 100%)`,
+        backgroundImage: `linear-gradient(180deg, rgba(24,34,78,1) ${!isFront && hideGradientBackground ? '100%' : '25%'
+            }, rgba(139,145,167,1) 100%)`,
         backgroundSize: 'cover',
         backgroundPosition: 'center',
         backgroundAttachment: 'fixed',
@@ -81,13 +79,14 @@ export const VCIDDisplayCard: React.FC<VCIDDisplayCardProps> = ({
         <Flipper className="w-full" flipKey={isFront}>
             <Flipped flipId="card">
                 <section
-                    className={`vc-display-card font-mouse flex flex-col items-center border-solid border-white rounded-[30px] z-10 max-w-[400px] relative bg-white shadow-3xl ${
-                        isFront ? '' : 'min-h-[800px]'
-                    }`}
+                    className={`vc-display-card font-mouse flex flex-col items-center border-solid border-white rounded-[30px] z-10 max-w-[400px] relative bg-white shadow-3xl ${isFront ? '' : 'min-h-[800px]'
+                        }`}
                 >
                     <div
                         className="relative vc-card-content-container flex flex-col items-center grow min-h-0 w-full rounded-t-[30px] rounded-b-[30px] overflow-hidden"
                         style={backgroundStyle}
+                        role="button"
+                        onClick={() => setIsFront(!isFront)}
                     >
                         <Flipped flipId="scroll-container">
                             <div className="vc-card-content-scroll-container w-full min-h-full flex flex-col justify-start items-center rounded-t-[30px] rounded-b-[30px]  scrollbar-hide pt-[20px]">