+++ packages/learn-card-init/CHANGELOG.md
@@ -1,5 +1,19 @@
 # learn-card-core
 
+## 1.2.28
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-plugin@1.7.11
+
+## 1.2.27
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-plugin@1.7.10
+
 ## 1.2.26
 
 ### Patch Changes